MFSCNet/requirements.txt
GDAL==3.0.4
matplotlib==3.2.2
numpy==1.19.2
opencv_python==4.5.3.56
Pillow==9.3.0
pycocotools==2.0.3
scikit_image==0.17.2
scikit_learn==1.1.3
scipy==1.5.4
skimage==0.0
torch==1.7.1
torchsummary==1.5.1
torchvision==0.8.2
tqdm==4.62.2
